# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Crux API
  version: 0.1.0
  summary: Backend contract for Crux v1 (Phases 2-5).
  description: |
    Source of truth for Crux HTTP contracts. Pydantic models under
    `backend/contracts/` are generated from this document via
    `backend/scripts/gen_contracts.py`. Edit YAML, then regen.

    See `docs/superpowers/specs/2026-05-03-crux-v1-design.md`
    sections 3.3, 3.4, 4.1-4.4, 5 for upstream design constraints.
servers:
  - url: http://localhost:8000
    description: Local development backend
security:
  - BearerAuth: []
...
paths:
...
  /api/upload:
    post:
      tags: [upload]
      summary: Upload a reference file (PDF, PPTX, TXT, or MD) to attach to a session (background ingestion).
      description: |
        Accepted file types are validated by extension: .pdf, .pptx, .txt, .md
        (.markdown is also accepted). The 25 MB per-request cap is unchanged.
        Ingestion is performed asynchronously by a background worker process;
        the document transitions pending -> processing -> ready (or failed)
        as the worker picks it up and completes extraction/embedding.
      operationId: uploadPdf
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [session_id, file]
              properties:
                session_id:
                  type: string
                  maxLength: 64
                file:
                  type: string
                  format: binary
      responses:
        "202":
          description: Accepted; ingestion is performed by a background worker process.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Session already ended (code session_ended).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          description: File content does not match its extension (code CONTENT_TYPE_MISMATCH).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
        "507":
          description: Storage write failed; the document row is marked failed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/upload/{document_id}:
    get:
      tags: [upload]
      summary: Poll ingestion status for an uploaded document.
      operationId: getUploadStatus
      parameters:
        - in: path
          name: document_id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Current ingestion status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadStatus"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase-issued JWT. Server validates against Supabase JWKS and
        resolves the request's `user_id` from the `sub` claim. Replaces the
        Phase 6-and-earlier convention of passing `user_id` via body/query.
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: >
        Rate limited. Error codes: daily_cap_reached, daily_cost_cap_reached,
        global_cost_cap_reached (daily caps, reset at midnight UTC), or
        too_many_requests (per-user burst limit; body carries retry_after_s
        and the Retry-After header is set).
      headers:
        Retry-After:
          description: Seconds until the burst window admits another request. Present only for too_many_requests.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PayloadTooLarge:
      description: Uploaded file exceeds the per-request size cap.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing, invalid, or expired bearer token.
    ServiceUnavailable:
      description: Upstream dependency unavailable (auth JWKS outage or LLM provider).
...
  schemas:
...
    IngestionStatus:
      type: string
      enum: [pending, processing, ready, failed]
...
    UploadResponse:
      type: object
      additionalProperties: false
      required: [document_id, session_id, filename, status]
      properties:
        document_id: { type: integer }
        session_id:  { type: string }
        filename:    { type: string }
        status:      { $ref: "#/components/schemas/IngestionStatus" }
    UploadStatus:
      type: object
      additionalProperties: false
      required: [id, status]
      description: Ingestion state for an uploaded document; polled by the UI.
      properties:
        id:     { type: integer }
        status: { $ref: "#/components/schemas/IngestionStatus" }
        error:  { type: [string, "null"], default: null }
...
    CodedErrorDetail:
      type: object
      additionalProperties: true
      required: [code]
      properties:
        code: { type: string }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - $ref: "#/components/schemas/CodedErrorDetail"

# ---------------------------------------------------------------------------
# SSE event contract reference (vendor extension — not processed by codegen)
# Describes the Server-Sent Events emitted by POST /chat/stream.
# Field names must match what agent/tutor.py actually yields.
# ---------------------------------------------------------------------------
...